Splitting up `SortableListItem`.
